PainterStylesMethods отдельным компонентом перенесено в папку API
